Extract tokenizer.js
